.kilo/worktrees/awake-koala/app/requirements.txt
datasets
apscheduler
PyGithub
firebase-admin
fastapi
uvicorn
requests
chromadb
